requirements.txt
fastapi
uvicorn 
pymongo[srv]
pyjwt
passlib[bcrypt]
